Change "player" field to "player_color" for clarity
The "player" field on the MatchTurn model was misleading as it did not represent the player object but the color used by the player. Therefore, it was renamed to "player_color" to improve code readability and avoid potential confusions in the future. This change was reflected in the relevant models, serializers, tests, and methods where the "player" field was previously referenced.